requirements.txt
discord.py
streamlit
pandas
python-dotenv
google-api-python-client
google-auth
google-auth-oauthlib
